Refactor json2csv module
